+++ packages/extension-koni-ui/src/assets/subwallet/index.ts
@@ -1,6 +1,7 @@
 // Copyright 2019-2022 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { SwapProviderId } from '@subwallet/extension-base/types/swap';
 import { DefaultLogosMap } from '@subwallet/extension-koni-ui/assets/logo';
 import { SUBSTRATE_GENERIC_KEY } from '@subwallet/extension-koni-ui/constants';
 
@@ -20,7 +21,10 @@ const SwLogosMap: Record<string, string> = {
   chain_flip_testnet: DefaultLogosMap.chain_flip,
   hydradx_mainnet: DefaultLogosMap.hydradx,
   hydradx_testnet: DefaultLogosMap.hydradx,
-  [SUBSTRATE_GENERIC_KEY]: DefaultLogosMap[SUBSTRATE_GENERIC_KEY]
+  [SUBSTRATE_GENERIC_KEY]: DefaultLogosMap[SUBSTRATE_GENERIC_KEY],
+  [SwapProviderId.POLKADOT_ASSET_HUB.toLowerCase()]: DefaultLogosMap.polkadot_assethub,
+  [SwapProviderId.KUSAMA_ASSET_HUB.toLowerCase()]: DefaultLogosMap.kusama_assethub,
+  [SwapProviderId.ROCOCO_ASSET_HUB.toLowerCase()]: DefaultLogosMap.rococo_assethub
 };
 
 export default SwLogosMap;