+++ packages/extension-web-ui/src/Popup/Transaction/variants/Swap.tsx
@@ -18,7 +18,7 @@ import { WebUIContext } from '@subwallet/extension-web-ui/contexts/WebUIContext'
 import { useChainConnection, useNotification, usePreCheckAction, useSelector, useTransactionContext, useWatchTransaction } from '@subwallet/extension-web-ui/hooks';
 import { getLatestSwapQuote, handleSwapRequest, handleSwapStep, validateSwapProcess } from '@subwallet/extension-web-ui/messaging/transaction/swap';
 import { FreeBalance, FreeBalanceToEarn, TransactionContent, TransactionFooter } from '@subwallet/extension-web-ui/Popup/Transaction/parts';
-import { DEFAULT_SWAP_PROCESS, SwapActionType, swapReducer } from '@subwallet/extension-web-ui/reducer';
+import { CommonActionType, commonProcessReducer, DEFAULT_COMMON_PROCESS } from '@subwallet/extension-web-ui/reducer';
 import { RootState } from '@subwallet/extension-web-ui/stores';
 import { Theme } from '@subwallet/extension-web-ui/themes';
 import { FormCallbacks, FormFieldData, SwapParams, ThemeProps, TokenSelectorItemType } from '@subwallet/extension-web-ui/types';
@@ -108,6 +108,7 @@ const Component = () => {
   const [submitLoading, setSubmitLoading] = useState(false);
   const [handleRequestLoading, setHandleRequestLoading] = useState(true);
   const [requestUserInteractToContinue, setRequestUserInteractToContinue] = useState<boolean>(false);
+  const [isButtonClicked, setIsButtonClicked] = useState(false);
   const continueRefreshQuoteRef = useRef<boolean>(false);
   const { token } = useTheme() as Theme;
 
@@ -144,7 +145,7 @@ const Component = () => {
   const { checkChainConnected, turnOnChain } = useChainConnection();
   const onPreCheck = usePreCheckAction(fromValue);
 
-  const [processState, dispatchProcessState] = useReducer(swapReducer, DEFAULT_SWAP_PROCESS);
+  const [processState, dispatchProcessState] = useReducer(commonProcessReducer, DEFAULT_COMMON_PROCESS);
 
   const fromAndToTokenMap = useMemo<Record<string, string[]>>(() => {
     const result: Record<string, string[]> = {};
@@ -309,6 +310,8 @@ const Component = () => {
   }, [form]);
 
   const onSwitchSide = useCallback(() => {
+    setIsButtonClicked(true);
+
     if (fromTokenSlugValue && toTokenSlugValue) {
       form.setFieldsValue({
         fromTokenSlug: toTokenSlugValue,
@@ -474,9 +477,8 @@ const Component = () => {
         type: 'error',
         duration: 8
       });
-
       dispatchProcessState({
-        type: SwapActionType.STEP_ERROR_ROLLBACK,
+        type: CommonActionType.STEP_ERROR_ROLLBACK,
         payload: error
       });
     },
@@ -510,7 +512,7 @@ const Component = () => {
             return false;
           } else {
             dispatchProcessState({
-              type: needRollback ? SwapActionType.STEP_ERROR_ROLLBACK : SwapActionType.STEP_ERROR,
+              type: needRollback ? CommonActionType.STEP_ERROR_ROLLBACK : CommonActionType.STEP_ERROR,
               payload: _errors[0]
             });
 
@@ -518,7 +520,7 @@ const Component = () => {
           }
         } else if (id) {
           dispatchProcessState({
-            type: SwapActionType.STEP_COMPLETE,
+            type: CommonActionType.STEP_COMPLETE,
             payload: rs
           });
 
@@ -576,7 +578,7 @@ const Component = () => {
 
       const submitData = async (step: number): Promise<boolean> => {
         dispatchProcessState({
-          type: SwapActionType.STEP_SUBMIT,
+          type: CommonActionType.STEP_SUBMIT,
           payload: null
         });
 
@@ -601,11 +603,11 @@ const Component = () => {
               return false;
             } else {
               dispatchProcessState({
-                type: SwapActionType.STEP_COMPLETE,
+                type: CommonActionType.STEP_COMPLETE,
                 payload: true
               });
               dispatchProcessState({
-                type: SwapActionType.STEP_SUBMIT,
+                type: CommonActionType.STEP_SUBMIT,
                 payload: null
               });
 
@@ -685,31 +687,19 @@ const Component = () => {
     }
   }, [accounts, chainValue, checkChainConnected, closeAlert, currentOptimalSwapPath, currentQuote, currentQuoteRequest, currentSlippage.slippage, notify, onError, onSuccess, openAlert, processState.currentStep, processState.steps.length, t]);
 
-  const destinationSwapValue = useMemo(() => {
-    if (currentQuote) {
-      const decimals = _getAssetDecimals(fromAssetInfo);
-
-      return new BigN(fromAmountValue || 0)
-        .div(BN_TEN.pow(decimals))
-        .multipliedBy(currentQuote.rate);
-    }
-
-    return BN_ZERO;
-  }, [currentQuote, fromAmountValue, fromAssetInfo]);
-
   const minimumReceived = useMemo(() => {
-    const calcMinimumReceived = (value: BigN) => {
+    const calcMinimumReceived = (value: string) => {
       const adjustedValue = supportSlippageSelection
         ? value
-        : value.multipliedBy(new BigN(1).minus(currentSlippage.slippage));
+        : new BigN(value).multipliedBy(new BigN(1).minus(currentSlippage.slippage)).integerValue(BigN.ROUND_DOWN);
 
       return adjustedValue.toString().includes('e')
         ? formatNumberString(adjustedValue.toString())
         : adjustedValue.toString();
     };
 
-    return calcMinimumReceived(destinationSwapValue);
-  }, [supportSlippageSelection, destinationSwapValue, currentSlippage.slippage]);
+    return calcMinimumReceived(currentQuote?.toAmount || '0');
+  }, [currentQuote?.toAmount, currentSlippage.slippage, supportSlippageSelection]);
 
   const onAfterConfirmTermModal = useCallback(() => {
     return setConfirmedTerm('swap-term-confirmed');
@@ -804,6 +794,12 @@ const Component = () => {
     return processState.steps.some((item) => item.type === CommonStepType.XCM);
   }, [processState.steps]);
 
+  const isSwapAssetHub = useMemo(() => {
+    const providerId = currentQuote?.provider?.id;
+
+    return providerId ? [SwapProviderId.KUSAMA_ASSET_HUB, SwapProviderId.POLKADOT_ASSET_HUB, SwapProviderId.ROCOCO_ASSET_HUB].includes(providerId) : false;
+  }, [currentQuote?.provider?.id]);
+
   const renderAlertBox = () => {
     const multichainAsset = fromAssetInfo?.multiChainAsset;
     const fromAssetName = multichainAsset && multiChainAssetMap[multichainAsset]?.name;
@@ -811,7 +807,15 @@ const Component = () => {
 
     return (
       <>
-        {isSwapXCM && fromAssetName && toAssetName && (
+        {isSwapAssetHub && !isFormInvalid && (
+          <AlertBox
+            className={'__assethub-notification'}
+            description={'Swapping on Asset Hub is in beta with a limited number of pairs and low liquidity. Continue at your own risk'}
+            title={'Pay attention!'}
+            type='warning'
+          />
+        )}
+        {isSwapXCM && fromAssetName && toAssetName && !isFormInvalid && (
           <AlertBox
             className={'__xcm-notification'}
             description={`The amount you entered is higher than your available balance on ${toAssetName} network. You need to first transfer cross-chain from ${fromAssetName} network to ${toAssetName} network to continue swapping`}
@@ -948,7 +952,7 @@ const Component = () => {
                   steps: result.process.steps,
                   feeStructure: result.process.totalFee
                 },
-                type: SwapActionType.STEP_CREATE
+                type: CommonActionType.STEP_CREATE
               });
 
               setQuoteOptions(result.quote.quotes);
@@ -1110,6 +1114,12 @@ const Component = () => {
     }
   }, [checkChainConnected, altChain, turnOnChain]);
 
+  useEffect(() => {
+    if (isButtonClicked) {
+      setIsButtonClicked(false);
+    }
+  }, [isButtonClicked]);
+
   const isNotConnectedAltChain = useMemo(() => {
     if (altChain && !checkChainConnected(altChain)) {
       return true;
@@ -1168,6 +1178,7 @@ const Component = () => {
                   <SwapFromField
                     amountValue={fromAmountValue}
                     fromAsset={fromAssetInfo}
+                    isButtonClicked={isButtonClicked}
                     label={t('From')}
                     onChangeAmount={onChangeAmount}
                     onSelectToken={onSelectFromToken}
@@ -1195,9 +1206,10 @@ const Component = () => {
                   </div>
 
                   <SwapToField
+                    decimals={_getAssetDecimals(toAssetInfo)}
                     loading={handleRequestLoading && showQuoteArea}
                     onSelectToken={onSelectToToken}
-                    swapValue={destinationSwapValue}
+                    swapValue={currentQuote?.toAmount || 0}
                     toAsset={toAssetInfo}
                     tokenSelectorItems={toTokenItems}
                     tokenSelectorValue={toTokenSlugValue}
@@ -1393,8 +1405,7 @@ const Component = () => {
                       shape='squircle'
                       size={24}
                     />
-
-                    {currentQuote.provider.name}
+                    <span className={'__provider-name'}>{currentQuote.provider.name}</span>
                   </MetaInfo.Default>
 
                   <MetaInfo.Default
@@ -1406,7 +1417,7 @@ const Component = () => {
                   <div className={'__minimum-received'}>
                     <MetaInfo.Number
                       customFormatter={swapCustomFormatter}
-                      decimals={0}
+                      decimals={_getAssetDecimals(toAssetInfo)}
                       formatType={'custom'}
                       label={
                         <Tooltip
@@ -1588,8 +1599,13 @@ const Swap = styled(Wrapper)<Props>(({ theme: { token } }: Props) => {
       alignItems: 'center',
       cursor: 'pointer'
     },
-    '.__xcm-notification': {
+    '.__xcm-notification, .__assethub-notification': {
       marginBottom: token.marginSM
+    },
+    '.__provider-name': {
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+      overflow: 'hidden'
     },
     '.__fee-paid-token': {
       display: 'flex',
@@ -1607,7 +1623,14 @@ const Swap = styled(Wrapper)<Props>(({ theme: { token } }: Props) => {
     },
     '.__swap-provider .__value ': {
       display: 'flex',
-      gap: 8
+      gap: 8,
+      overflow: 'hidden'
+    },
+    '.__swap-provider .__col': {
+      alignItems: 'unset',
+      flexDirection: 'row',
+      justifyContent: 'flex-end'
+
     },
     '.ant-background-icon': {
       width: 24,